moved Data and cleaned up
