+++ legacy/modem.py
@@ -472,7 +472,7 @@ class Data(object):
         >>> edi_list = [os.path.join(edi_path, edi) \
                         for edi in os.listdir(edi_path)\
                         if edi.find('.edi') > 0]
-        >>> md = modem.Data(edi_list, period_min=.1, period_max=300,\
+import mtpy.modeling.ModEM        >>> md = mtpy.modeling.ModEM.Data(edi_list, period_min=.1, period_max=300,\
                             max_num_periods=12)
         >>> md.write_data_file(save_path=r"/home/modem/inv1")
         
@@ -484,7 +484,7 @@ class Data(object):
         >>> edi_list = [os.path.join(edi_path, edi) \
                         for edi in os.listdir(edi_path)\
                         if edi.find('.edi') > 0]
-        >>> md = modem.Data(edi_list)
+import mtpy.modeling.ModEM        >>> md = mtpy.modeling.ModEM.Data(edi_list)
         >>> #get period list from an .edi file
         >>> mt_obj1 = modem.mt.MT(edi_list[0])
         >>> inv_period_list = 1./mt_obj1.Z.freq
@@ -495,8 +495,8 @@ class Data(object):
                 
     :Example 3 --> change error values: ::
         
-        >>> import mtpy.modeling.modem as modem
-        >>> mdr = modem.Data()
+import mtpy.modeling.ModEM        >>> import mtpy.modeling.modem as modem
+        >>> mdr = mtpy.modeling.ModEM.Data()
         >>> mdr.read_data_file(r"/home/modem/inv1/ModEM_Data.dat")
         >>> mdr.error_type = 'floor'
         >>> mdr.error_floor = 10
@@ -505,8 +505,8 @@ class Data(object):
         
     :Example 4 --> change inversion type: ::
         
-        >>> import mtpy.modeling.modem as modem
-        >>> mdr = modem.Data()
+import mtpy.modeling.ModEM        >>> import mtpy.modeling.modem as modem
+        >>> mdr = mtpy.modeling.ModEM.Data()
         >>> mdr.read_data_file(r"/home/modem/inv1/ModEM_Data.dat")
         >>> mdr.inv_mode = '3'
         >>> mdr.write_data_file(save_path=r"/home/modem/inv2")
@@ -1017,7 +1017,7 @@ class Data(object):
             >>> edi_list = [os.path.join(edi_path, edi) \
                             for edi in os.listdir(edi_path)\
                             if edi.find('.edi') > 0]
-            >>> md = modem.Data(edi_list, period_min=.1, period_max=300,\
+import mtpy.modeling.ModEM            >>> md = mtpy.modeling.ModEM.Data(edi_list, period_min=.1, period_max=300,\
                                 max_num_periods=12)
             >>> md.write_data_file(save_path=r"/home/modem/inv1")
         """
@@ -1172,8 +1172,8 @@ class Data(object):
         
         :Example: ::
             
-            >>> import mtpy.modeling.modem as modem
-            >>> mdr = modem.Data()
+import mtpy.modeling.ModEM            >>> import mtpy.modeling.modem as modem
+            >>> mdr = mtpy.modeling.ModEM.Data()
             >>> mdr.convert_ws3dinv_data_file(r"/home/ws3dinv/inv1/WSData.dat",
                     station_fn=r"/home/ws3dinv/inv1/WS_Station_Locations.txt")
         """
@@ -1257,8 +1257,8 @@ class Data(object):
         Example
         -----------
             :Convert ModEM data file to WS: ::
-                >>> import mtpy.modeling.modem as modem
-                >>> md = modem.Data()
+import mtpy.modeling.ModEM                >>> import mtpy.modeling.modem as modem
+                >>> md = mtpy.modeling.ModEM.Data()
                 >>> md.convert_modem_to_ws(data_fn=r"/home/mt/modem/data.dat")
         """
         
@@ -1513,8 +1513,8 @@ class Data(object):
             
             #self.mt_dict[s_key].zinv.compute_invariants()
             self.mt_dict[s_key].pt.set_z_object(mt_obj.Z)
-            self.mt_dict[s_key].Tipper._compute_amp_phase()
-            self.mt_dict[s_key].Tipper._compute_mag_direction()
+            self.mt_dict[s_key].Tipper.compute_amp_phase()
+            self.mt_dict[s_key].Tipper.compute_mag_direction()
                                    
         
             self.data_array[ii]['station'] = mt_obj.station
