+++ mtpy/gui/modem_plot_pt_maps_qt5.py
@@ -14,6 +14,8 @@ JP 2014
 import os
 import numpy as np
 
+import mtpy.modeling.ModEM
+
 try:
     from PyQt5 import QtCore, QtGui, QtWidgets
 except ImportError:
@@ -309,7 +311,7 @@ class Ui_MainWindow(mtplottools.MTArrows, mtplottools.MTEllipse):
         
         fn = os.path.abspath(fn)
                                        
-        self.modem_data = modem.Data()
+        self.modem_data = mtpy.modeling.ModEM.Data()
         self.modem_data.read_data_file(fn)
         self.modem_data_fn = fn
         
@@ -365,7 +367,7 @@ class Ui_MainWindow(mtplottools.MTArrows, mtplottools.MTEllipse):
                                            filter='*.dat', 
                                            directory=self.dir_path))
                                        
-        self.modem_resp = modem.Data()
+        self.modem_resp = mtpy.modeling.ModEM.Data()
         self.modem_resp.read_data_file(fn)
         self.modem_resp_fn = fn
         self.plot()
@@ -452,7 +454,7 @@ class Ui_MainWindow(mtplottools.MTArrows, mtplottools.MTEllipse):
 
             # compute tipper data
             tip = self.modem_data.mt_dict[key].Tipper
-            tip._compute_mag_direction()
+            tip.compute_mag_direction()
                         
             data_pt_arr[:, ii]['txr'] = tip.mag_real*\
                                         np.sin(np.deg2rad(tip.angle_real))
@@ -473,7 +475,7 @@ class Ui_MainWindow(mtplottools.MTArrows, mtplottools.MTEllipse):
                 model_pt_arr[:, ii]['skew'] = mpt.beta[0]
                 
                 mtip = self.modem_resp.mt_dict[key].Tipper
-                mtip._compute_mag_direction()
+                mtip.compute_mag_direction()
                             
                 model_pt_arr[:, ii]['txr'] = mtip.mag_real*\
                                             np.sin(np.deg2rad(mtip.angle_real))
