+++ mtpy/modeling/ModEM/__init__.py
@@ -1,1 +1,5 @@
-from exception import ModEMError
+from .exception import ModEMError
+from .station import Stations
+from .data import Data
+
+__all__ = ['ModEMError', 'Stations', 'Data']
