+++ mtpy/modeling/modem.py
@@ -31,9 +31,7 @@ from matplotlib.patches import Ellipse
 from matplotlib.ticker import MultipleLocator, FormatStrFormatter
 
 import mtpy.analysis.pt as mtpt
-import mtpy.core.mt as mt
 # mtpy modules
-import mtpy.core.z as mtz
 import mtpy.imaging.mtcolors as mtcl
 import mtpy.imaging.mtplottools as mtplottools
 import mtpy.modeling.ws3dinv as ws
@@ -43,7 +41,7 @@ import mtpy.utils.filehandling as mtfh
 import mtpy.utils.gis_tools as gis_tools
 import mtpy.utils.mesh_tools as mtmesh
 # vtk tools
-from mtpy.modeling.ModEM import ModEMError
+from mtpy.modeling.ModEM import ModEMError, Data
 
 try:
     from evtk.hl import gridToVTK, pointsToVTK
@@ -53,1667 +51,6 @@ except ImportError:
 
 
 # ==============================================================================
-class Stations(object):
-    """
-    station locations class
-    
-    ..note:: If the survey steps across multiple UTM zones, then a 
-             distance will be added to the stations to place them in 
-             the correct location.  This distance is 
-             _utm_grid_size_north and _utm_grid_size_east.  You should 
-             these parameters to place the locations in the proper spot
-             as grid distances and overlaps change over the globe.
-             **This is not implemented yet**
-    """
-
-    def __init__(self, **kwargs):
-
-        self.dtype = [('station', '|S10'),
-                      ('lat', np.float),
-                      ('lon', np.float),
-                      ('elev', np.float),
-                      ('rel_east', np.float),
-                      ('rel_north', np.float),
-                      ('east', np.float),
-                      ('north', np.float),
-                      ('zone', 'S4')]
-        self.station_locations = np.zeros(0, dtype=self.dtype)
-        self.model_epsg = None
-        self.model_utm_zone = None
-
-        for key in kwargs.keys():
-            if hasattr(self, key):
-                setattr(self, key, kwargs[key])
-
-    ## --> define properties that can only be returned and not set    
-    @property
-    def lat(self):
-        return self.station_locations['lat']
-
-    @property
-    def lon(self):
-        return self.station_locations['lon']
-
-    @property
-    def east(self):
-        return self.station_locations['east']
-
-    @property
-    def north(self):
-        return self.station_locations['north']
-
-    @property
-    def elev(self):
-        return self.station_locations['elev']
-
-    @property
-    def rel_east(self):
-        return self.station_locations['rel_east']
-
-    @property
-    def rel_north(self):
-        return self.station_locations['rel_north']
-
-    @property
-    def utm_zone(self):
-        return self.station_locations['zone']
-
-    @property
-    def station(self):
-        return self.station_locations['station']
-
-    def _get_mt_objs_from_list(self, input_list):
-        """
-        get mt_objects from a list of files or mt_objects
-        """
-
-        if type(input_list) not in [list, np.ndarray]:
-            raise ValueError('Input list needs to be type list, not {0}'.format(type(input_list)))
-
-        if type(input_list[0]) is mt.MT:
-            return input_list
-
-        if type(input_list[0]) is str:
-            if input_list[0].endswith('.edi'):
-                return [mt.MT(fn) for fn in input_list]
-
-            else:
-                raise ModEMError('file {0} not supported yet'.format(input_list[0][-4:]))
-
-    def get_station_locations(self, input_list):
-        """
-        get station locations from a list of edi files
-        
-        Arguments
-        -------------
-            **input_list** : list
-                             list of edi file names, or mt_objects
-                             
-        
-        Returns
-        ------------
-            * fills station_locations array
-            
-        """
-
-        mt_obj_list = self._get_mt_objs_from_list(input_list)
-
-        # if station locations are not input read from the edi files
-        if mt_obj_list is None:
-            raise AttributeError('mt_obj_list is None, need to input a list of '
-                                 'mt objects to read in.')
-
-        n_stations = len(mt_obj_list)
-
-        if n_stations == 0:
-            raise ModEMError('No .edi files in edi_list, please check '
-                             'file locations.')
-
-        # make a structured array to put station location information into
-        self.station_locations = np.zeros(n_stations,
-                                          dtype=self.dtype)
-        # get station locations in meters
-        for ii, mt_obj in enumerate(mt_obj_list):
-            self.station_locations[ii]['lat'] = mt_obj.lat
-            self.station_locations[ii]['lon'] = mt_obj.lon
-            self.station_locations[ii]['station'] = mt_obj.station
-            self.station_locations[ii]['elev'] = mt_obj.elev
-
-            if ((self.model_epsg is not None) or (self.model_utm_zone is not None)):
-                east, north, utm_zone = gis_tools.project_point_ll2utm(mt_obj.lat,
-                                                                       mt_obj.lon,
-                                                                       utm_zone=self.model_utm_zone,
-                                                                       epsg=self.model_epsg)
-                self.station_locations[ii]['east'] = east
-                self.station_locations[ii]['north'] = north
-                self.station_locations[ii]['zone'] = utm_zone
-            else:
-                self.station_locations[ii]['east'] = mt_obj.east
-                self.station_locations[ii]['north'] = mt_obj.north
-                self.station_locations[ii]['zone'] = mt_obj.utm_zone
-
-        # get relative station locations
-        self.calculate_rel_locations()
-
-    def calculate_rel_locations(self, shift_east=0, shift_north=0):
-        """
-        put station in a coordinate system relative to 
-        (shift_east, shift_north)
-        (+) shift right or up
-        (-) shift left or down
-        
-        """
-        #
-        #        #remove the average distance to get coordinates in a relative space
-        #        self.station_locations['rel_east'] = self.east-self.east.mean()
-        #        self.station_locations['rel_north'] = self.north-self.north.mean()
-        #
-        #        #translate the stations so they are relative to 0,0
-        #        east_center = (self.rel_east.max()-np.abs(self.rel_east.min()))/2.
-        #        north_center = (self.rel_north.max()-np.abs(self.rel_north.min()))/2.
-        #
-        #
-        #        #remove the average distance to get coordinates in a relative space
-        #        self.station_locations['rel_east'] -= east_center+shift_east
-        #        self.station_locations['rel_north'] -= north_center+shift_north
-
-        # translate the stations so they are relative to 0,0
-        east_center = (self.east.max() + self.east.min()) / 2.
-        north_center = (self.north.max() + self.north.min()) / 2.
-
-        self.station_locations['rel_east'] = self.east - east_center
-        self.station_locations['rel_north'] = self.north - north_center
-
-    # make center point a get method, can't set it.
-    @property
-    def center_point(self):
-        """
-        calculate the center point from the given station locations
-        
-        Returns
-        -------------
-            **center_location** : np.ndarray
-                                  structured array of length 1
-                                  dtype includes (east, north, zone, lat, lon)
-        """
-        center_location = np.recarray(1, dtype=self.dtype)
-        #        AK - using the mean here but in get_relative_locations used (max + min)/2, why???
-
-        #        center_point = np.array([self.east.mean(), self.north.mean()])
-        #
-        #        #translate the stations so they are relative to 0,0
-        #        east_center = (self.rel_east.max()-np.abs(self.rel_east.min()))/2
-        #        north_center = (self.rel_north.max()-np.abs(self.rel_north.min()))/2
-        #
-        #        center_point[0] -= east_center
-        #        center_point[1] -= north_center
-        #
-        #        # calculate center point in lat, lon, easting, northing
-        #        center_location['east'] = center_point[0]
-        #        center_location['north'] = center_point[1]
-
-        center_point = np.array([self.east.max() + self.east.min(),
-                                 self.north.max() + self.north.min()]) / 2.
-        center_location['east'] = center_point[0]
-        center_location['north'] = center_point[1]
-
-        center_location['zone'] = self.utm_zone[0]
-
-        center_ll = gis_tools.project_point_utm2ll(float(center_point[0]),
-                                                   float(center_point[1]),
-                                                   self.utm_zone[0],
-                                                   epsg=self.model_epsg)
-
-        center_location['lat'] = center_ll[0]
-        center_location['lon'] = center_ll[1]
-
-        return center_location
-
-    def rotate_stations(self, rotation_angle):
-        """
-        Rotate stations assuming N is 0
-        
-        Arguments
-        -------------
-            **rotation_angle** : float
-                                 angle in degrees assuming N is 0
-        
-        Returns
-        -------------
-            * refils rel_east and rel_north in station_locations.  Does this
-              because you will still need the original locations for plotting
-              later.
-              
-        """
-
-        cos_ang = np.cos(np.deg2rad(rotation_angle))
-        sin_ang = np.sin(np.deg2rad(rotation_angle))
-        rot_matrix = np.matrix(np.array([[cos_ang, sin_ang],
-                                         [-sin_ang, cos_ang]]))
-
-        coords = np.array([self.station_locations['rel_east'],
-                           self.station_locations['rel_north']])
-
-        # rotate the relative station locations
-        new_coords = np.array(np.dot(rot_matrix, coords))
-
-        self.station_locations['rel_east'][:] = new_coords[0, :]
-        self.station_locations['rel_north'][:] = new_coords[1, :]
-
-        print 'Rotated stations by {0:.1f} deg clockwise from N'.format(
-            rotation_angle)
-
-    def check_utm_crossing(self):
-        """
-        If the stations cross utm zones, then estimate distance by computing
-        distance on a sphere.
-        """
-        #
-        #        latMid = (Lat1+Lat2 )/2.0;  // or just use Lat1 for slightly less accurate estimate
-        #
-        #
-        #        m_per_deg_lat = 111132.954 - 559.822 * cos( 2.0 * latMid ) + 1.175 * cos( 4.0 * latMid);
-        #        m_per_deg_lon = (3.14159265359/180 ) * 6367449 * cos ( latMid );
-        #
-        #        deltaLat = fabs(Lat1 - Lat2);
-        #        deltaLon = fabs(Lon1 - Lon2);
-        #
-        #        dist_m = sqrt (  pow( deltaLat * m_per_deg_lat,2) + pow( deltaLon * m_per_deg_lon , 2) );
-        #
-        pass
-
-
-class Data(object):
-    """
-    Data will read and write .dat files for ModEM and convert a WS data file 
-    to ModEM format.
-    
-    ..note: :: the data is interpolated onto the given periods such that all
-               stations invert for the same periods.  The interpolation is
-               a linear interpolation of each of the real and imaginary parts 
-               of the impedance tensor and induction tensor.    
-               See mtpy.core.mt.MT.interpolate for more details
-    
-    Arguments
-    ------------
-        **edi_list** : list 
-                       list of full paths to .edi files you want to invert for
-        
-    ====================== ====================================================
-    Attributes/Key Words   Description    
-    ====================== ====================================================
-    _dtype                 internal variable defining the data type of 
-                           data_array 
-    _t_shape               internal variable defining shape of tipper array in
-                           _dtype  
-    _z_shape               internal variable defining shape of Z array in 
-                           _dtype
-    center_position        (east, north, evel) for center point of station 
-                           array.  All stations are relative to this location
-                           for plotting purposes.
-    comp_index_dict        dictionary for index values of component of Z and T
-    station_locations      Stations object
-    data_array             numpy.ndarray (num_stations) structured to store
-                           data.  keys are:
-                               * station --> station name
-                               * lat --> latitude in decimal degrees
-                               * lon --> longitude in decimal degrees
-                               * elev --> elevation (m)
-                               * rel_east -- > relative east location to 
-                                               center_position (m)
-                               * rel_north --> relative north location to 
-                                               center_position (m)
-                               * east --> UTM east (m)
-                               * north --> UTM north (m)
-                               * zone --> UTM zone
-                               * z --> impedance tensor array with shape
-                                       (num_freq, 2, 2)
-                               * z_err --> impedance tensor error array with
-                                       shape (num_freq, 2, 2)
-                               * tip --> Tipper array with shape
-                                       (num_freq, 1, 2)
-                               * tipperr --> Tipper array with shape
-                                       (num_freq, 1, 2)
-    data_fn                full path to data file 
-    data_period_list       period list from all the data
-    edi_list               list of full paths to edi files
-    error_type_tipper      [ 'abs' | 'floor' ] 
-                           *default* is 'abs'
-    error_type_z           [ 'egbert' | 'mean_od' | 'eigen' ]
-                           *default* is 'egbert_floor'
-                                * add '_floor' to any of the above to set the
-                                  error as an error floor, otherwise all 
-                                  components are give weighted the same
-
-                                * 'egbert' sets error to  
-                                           error_value_z * sqrt(abs(zxy*zyx))
-                                * 'mean_od' sets error to 
-                                           error_value_z * mean([Zxy, Zyx])
-                                * 'eigen' sets error to
-                                          error_value_z * eigenvalues(Z[ii])
-
-                                           
-    error_value_z            percentage to multiply Z by to set error
-                           *default* is 5 for 5% of Z as error
-    error_value_tipper     absolute error between 0 and 1.
-    fn_basename            basename of data file. *default* is 'ModEM_Data.dat'
-    header_strings         strings for header of data file following the format
-                           outlined in the ModEM documentation
-    inv_comp_dict          dictionary of inversion componets
-    inv_mode               inversion mode, options are: *default* is '1'
-                               * '1' --> for 'Full_Impedance' and 
-                                             'Full_Vertical_Components'
-                               * '2' --> 'Full_Impedance'
-                               * '3' --> 'Off_Diagonal_Impedance' and 
-                                         'Full_Vertical_Components'
-                               * '4' --> 'Off_Diagonal_Impedance'
-                               * '5' --> 'Full_Vertical_Components'
-                               * '6' --> 'Full_Interstation_TF'
-                               * '7' --> 'Off_Diagonal_Rho_Phase' 
-
-    inv_mode_dict          dictionary for inversion modes
-    max_num_periods        maximum number of periods
-    model_epsg             epsg code for model projection, provide this to
-                           project model to non-utm coordinates. Find the epsg
-                           code for your projection on 
-                           http://spatialreference.org/ref/ or google search
-                           epsg "your projection"
-    model_utm_zone         alternative to model_epsg, choose a utm zone to
-                           project all sites to (e.g. '55S')
-    mt_dict                dictionary of mtpy.core.mt.MT objects with keys 
-                           being station names
-    period_dict            dictionary of period index for period_list
-    period_list            list of periods to invert for
-    period_max             maximum value of period to invert for
-    period_min             minimum value of period to invert for
-    rotate_angle           Angle to rotate data to assuming 0 is N and E is 90            
-    save_path              path to save data file to
-    units                  [ [V/m]/[T] | [mV/km]/[nT] | Ohm ] units of Z
-                           *default* is [mV/km]/[nT]
-    wave_sign_impedance    [ + | - ] sign of time dependent wave.  
-                           *default* is '+' as positive downwards. 
-    wave_sign_tipper       [ + | - ] sign of time dependent wave.  
-                           *default* is '+' as positive downwards. 
-    ====================== ====================================================
-   
-    ========================== ================================================
-    Methods                    Description 
-    ========================== ================================================
-    convert_ws3dinv_data_file  convert a ws3dinv file to ModEM fomrat, 
-                               **Note** this doesn't include tipper data and 
-                               you need a station location file like the one
-                               output by mtpy.modeling.ws3dinv
-    get_data_from_edi          get data from given .edi files and fill 
-                               attributes accordingly
-    get_mt_dict                get a dictionary of mtpy.core.mt.MT objects 
-                               with keys being station names
-    get_period_list            get a list of periods to invert for
-    get_station_locations      get station locations and relative locations
-                               filling in station_locations
-    read_data_file             read in a ModEM data file and fill attributes
-                               data_array, station_locations, period_list, mt_dict
-    write_data_file            write a ModEM data file 
-    ========================== ================================================
-    
-    
-    :Example 1 --> create inversion period list: ::
-    
-        >>> import os
-        >>> import mtpy.modeling.modem as modem
-        >>> edi_path = r"/home/mt/edi_files"
-        >>> edi_list = [os.path.join(edi_path, edi) \
-                        for edi in os.listdir(edi_path)\
-                        if edi.find('.edi') > 0]
-        >>> md = modem.Data(edi_list, period_min=.1, period_max=300,\
-                            max_num_periods=12)
-        >>> md.write_data_file(save_path=r"/home/modem/inv1")
-        
-    :Example 2 --> set inverions period list from data: ::
-    
-        >>> import os
-        >>> import mtpy.modeling.modem as modem
-        >>> edi_path = r"/home/mt/edi_files"
-        >>> edi_list = [os.path.join(edi_path, edi) \
-                        for edi in os.listdir(edi_path)\
-                        if edi.find('.edi') > 0]
-        >>> md = modem.Data(edi_list)
-        >>> #get period list from an .edi file
-        >>> mt_obj1 = modem.mt.MT(edi_list[0])
-        >>> inv_period_list = 1./mt_obj1.Z.freq
-        >>> #invert for every third period in inv_period_list
-        >>> inv_period_list = inv_period_list[np.arange(0, len(inv_period_list, 3))]
-        >>> md.period_list = inv_period_list        
-        >>> md.write_data_file(save_path=r"/home/modem/inv1")
-                
-    :Example 3 --> change error values: ::
-        
-        >>> import mtpy.modeling.modem as modem
-        >>> mdr = modem.Data()
-        >>> mdr.read_data_file(r"/home/modem/inv1/ModEM_Data.dat")
-        >>> mdr.error_type = 'floor'
-        >>> mdr.error_floor = 10
-        >>> mdr.error_tipper = .03
-        >>> mdr.write_data_file(save_path=r"/home/modem/inv2")
-        
-    :Example 4 --> change inversion type: ::
-        
-        >>> import mtpy.modeling.modem as modem
-        >>> mdr = modem.Data()
-        >>> mdr.read_data_file(r"/home/modem/inv1/ModEM_Data.dat")
-        >>> mdr.inv_mode = '3'
-        >>> mdr.write_data_file(save_path=r"/home/modem/inv2")
-        
-    :Example 5 --> rotate data: ::
-        
-        >>> md.rotation_angle = 60
-        >>> md.write_data_file(save_path=r"/home/modem/Inv1")
-        >>> # or
-        >>> md.write_data_file(save_path=r"/home/modem/Inv1", \
-                               rotation_angle=60)
-    
-                  
-    """
-
-    def __init__(self, edi_list=None, **kwargs):
-        self.edi_list = edi_list
-
-        self.error_type_z = 'egbert_floor'
-        self.error_value_z = 5.0
-
-        self.error_value_tipper = .05
-        self.error_type_tipper = 'abs'
-
-        self.wave_sign_impedance = '+'
-        self.wave_sign_tipper = '+'
-        self.units = '[mV/km]/[nT]'
-        self.inv_mode = '1'
-
-        self.period_list = None
-        self.period_step = 1
-        self.period_min = None
-        self.period_max = None
-        self.period_buffer = None
-        self.max_num_periods = None
-        self.data_period_list = None
-
-        self.data_fn = 'ModEM_Data.dat'
-        self.save_path = os.getcwd()
-
-        self.formatting = '1'
-
-        self._rotation_angle = 0.0
-        self._set_rotation_angle(self._rotation_angle)
-
-        self.center_point = None
-
-        self.data_array = None
-        self.mt_dict = None
-        self.model_utm_zone = None
-        self.model_epsg = None
-
-        self._z_shape = (1, 2, 2)
-        self._t_shape = (1, 1, 2)
-        self._dtype = [('station', '|S10'),
-                       ('lat', np.float),
-                       ('lon', np.float),
-                       ('elev', np.float),
-                       ('rel_east', np.float),
-                       ('rel_north', np.float),
-                       ('east', np.float),
-                       ('north', np.float),
-                       ('zone', '|S4'),
-                       ('z', (np.complex, self._z_shape)),
-                       ('z_err', (np.float, self._z_shape)),
-                       ('z_inv_err', (np.float, self._z_shape)),
-                       ('tip', (np.complex, self._t_shape)),
-                       ('tip_err', (np.float, self._t_shape)),
-                       ('tip_inv_err', (np.float, self._t_shape))]
-
-        self.inv_mode_dict = {'1': ['Full_Impedance', 'Full_Vertical_Components'],
-                              '2': ['Full_Impedance'],
-                              '3': ['Off_Diagonal_Impedance',
-                                    'Full_Vertical_Components'],
-                              '4': ['Off_Diagonal_Impedance'],
-                              '5': ['Full_Vertical_Components'],
-                              '6': ['Full_Interstation_TF'],
-                              '7': ['Off_Diagonal_Rho_Phase']}
-        self.inv_comp_dict = {'Full_Impedance': ['zxx', 'zxy', 'zyx', 'zyy'],
-                              'Off_Diagonal_Impedance': ['zxy', 'zyx'],
-                              'Full_Vertical_Components': ['tx', 'ty']}
-
-        self.comp_index_dict = {'zxx': (0, 0), 'zxy': (0, 1), 'zyx': (1, 0),
-                                'zyy': (1, 1), 'tx': (0, 0), 'ty': (0, 1)}
-
-        self.header_string = ' '.join(['# Period(s)',
-                                       'Code',
-                                       'GG_Lat',
-                                       'GG_Lon',
-                                       'X(m)',
-                                       'Y(m)',
-                                       'Z(m)',
-                                       'Component',
-                                       'Real',
-                                       'Imag',
-                                       'Error\n'])
-
-        for key in kwargs.keys():
-            setattr(self, key, kwargs[key])
-
-    def _set_dtype(self, z_shape, t_shape):
-        """
-        reset dtype
-        """
-
-        self._z_shape = z_shape
-        self._t_shape = t_shape
-
-        self._dtype = [('station', '|S10'),
-                       ('lat', np.float),
-                       ('lon', np.float),
-                       ('elev', np.float),
-                       ('rel_east', np.float),
-                       ('rel_north', np.float),
-                       ('east', np.float),
-                       ('north', np.float),
-                       ('zone', '|S4'),
-                       ('z', (np.complex, self._z_shape)),
-                       ('z_err', (np.float, self._z_shape)),
-                       ('z_inv_err', (np.float, self._z_shape)),
-                       ('tip', (np.complex, self._t_shape)),
-                       ('tip_err', (np.float, self._t_shape)),
-                       ('tip_inv_err', (np.float, self._t_shape))]
-
-    def get_header_string(self, error_type, error_value, rotation_angle):
-        """
-        reset the header sring for file
-        """
-
-        h_str = ','.join(['# Created using MTpy calculated {0} error of {1:.0f}%',
-                          ' data rotated {2:.1f}_deg clockwise from N\n'])
-
-        return h_str.format(error_type, error_value, rotation_angle)
-
-    def get_mt_dict(self):
-        """
-        get mt_dict from edi file list
-        """
-
-        if self.edi_list is None:
-            raise ModEMError('edi_list is None, please input a list of '
-                             '.edi files containing the full path')
-
-        if len(self.edi_list) == 0:
-            raise ModEMError('edi_list is empty, please input a list of '
-                             '.edi files containing the full path')
-
-        self.mt_dict = {}
-        for edi in self.edi_list:
-            mt_obj = mt.MT(edi)
-            self.mt_dict[mt_obj.station] = mt_obj
-
-    def get_relative_station_locations(self):
-        """
-        get station locations from edi files
-        """
-
-        stations_obj = Stations(model_epsg=self.model_epsg,
-                                model_utm_zone=self.model_utm_zone)
-        mt_list = [self.mt_dict[s_key] for s_key in sorted(self.mt_dict.keys())]
-
-        stations_obj.get_station_locations(mt_list)
-        # rotate locations if needed
-        if self._rotation_angle != 0:
-            stations_obj.rotate_stations(self._rotation_angle)
-
-        # fill data array
-        self.data_array[:]['station'] = stations_obj.station
-        self.data_array[:]['lat'] = stations_obj.lat
-        self.data_array[:]['lon'] = stations_obj.lon
-        self.data_array[:]['east'] = stations_obj.east
-        self.data_array[:]['north'] = stations_obj.north
-        self.data_array[:]['elev'] = stations_obj.elev
-        self.data_array[:]['rel_east'] = stations_obj.rel_east
-        self.data_array[:]['rel_north'] = stations_obj.rel_north
-        self.data_array[:]['zone'] = stations_obj.utm_zone
-
-        # get center point
-        self.center_point = stations_obj.center_point
-
-    def get_period_list(self):
-        """
-        make a period list to invert for
-        
-        """
-        if self.mt_dict is None:
-            self.get_mt_dict()
-
-        if self.period_list is not None:
-            print '-' * 50
-            print 'Inverting for periods:'
-            for per in self.period_list:
-                print '     {0:<12.6f}'.format(per)
-            print '-' * 50
-            return
-
-        data_period_list = []
-        for s_key in sorted(self.mt_dict.keys()):
-            mt_obj = self.mt_dict[s_key]
-            data_period_list.extend(list(1. / mt_obj.Z.freq))
-
-        self.data_period_list = np.array(sorted(list(set(data_period_list)),
-                                                reverse=False))
-
-        if self.period_min is not None:
-            if self.period_max is None:
-                raise ModEMError('Need to input period_max')
-        if self.period_max is not None:
-            if self.period_min is None:
-                raise ModEMError('Need to input period_min')
-        if self.period_min is not None and self.period_max is not None:
-            if self.max_num_periods is None:
-                raise ModEMError('Need to input number of periods to use')
-
-            min_index = np.where(self.data_period_list >= self.period_min)[0][0]
-            max_index = np.where(self.data_period_list <= self.period_max)[0][-1]
-
-            pmin = np.log10(self.data_period_list[min_index])
-            pmax = np.log10(self.data_period_list[max_index])
-            self.period_list = np.logspace(pmin, pmax, num=self.max_num_periods)
-
-            print '-' * 50
-            print 'Inverting for periods:'
-            for per in self.period_list:
-                print '     {0:<12.6f}'.format(per)
-            print '-' * 50
-
-        if self.period_list is None:
-            raise ModEMError('Need to input period_min, period_max, '
-                             'max_num_periods or a period_list')
-
-    def _set_rotation_angle(self, rotation_angle):
-        """
-        on set rotation angle rotate mt_dict and data_array, 
-        """
-        if self._rotation_angle == rotation_angle:
-            return
-
-        print 'Changing rotation angle from {0:.1f} to {1:.1f}'.format(
-            self._rotation_angle, rotation_angle)
-
-        self._rotation_angle = -self._rotation_angle + rotation_angle
-
-        if self.rotation_angle == 0:
-            return
-
-        print 'Changing rotation angle from {0:.1f} to {1:.1f}'.format(
-            self._rotation_angle, rotation_angle)
-        self._rotation_angle = rotation_angle
-
-        if self.mt_dict is None:
-            self.get_mt_dict()
-
-        for mt_key in sorted(self.mt_dict.keys()):
-            mt_obj = self.mt_dict[mt_key]
-            mt_obj.Z.rotate(self._rotation_angle)
-            mt_obj.Tipper.rotate(self._rotation_angle)
-
-        print 'Data rotated to align with {0:.1f} deg clockwise from N'.format(
-            self._rotation_angle)
-        self.fill_data_array()
-
-    def _get_rotation_angle(self):
-        return self._rotation_angle
-
-    rotation_angle = property(fget=_get_rotation_angle,
-                              fset=_set_rotation_angle,
-                              doc="""Rotate data assuming N=0, E=90""")
-
-    def fill_data_array(self):
-        """
-        fill the data array from mt_dict
-        
-        """
-
-        if self.period_list is None:
-            self.get_period_list()
-
-        ns = len(self.mt_dict.keys())
-        nf = len(self.period_list)
-
-        d_array = False
-        if self.data_array is not None:
-            d_arr_copy = self.data_array.copy()
-            d_array = True
-
-        self._set_dtype((nf, 2, 2), (nf, 1, 2))
-        self.data_array = np.zeros(ns, dtype=self._dtype)
-
-        rel_distance = True
-        for ii, s_key in enumerate(sorted(self.mt_dict.keys())):
-            mt_obj = self.mt_dict[s_key]
-            if d_array is True:
-                try:
-                    d_index = np.where(d_arr_copy['station'] == s_key)[0][0]
-                    self.data_array[ii]['station'] = s_key
-                    self.data_array[ii]['lat'] = d_arr_copy[d_index]['lat']
-                    self.data_array[ii]['lon'] = d_arr_copy[d_index]['lon']
-                    self.data_array[ii]['east'] = d_arr_copy[d_index]['east']
-                    self.data_array[ii]['north'] = d_arr_copy[d_index]['north']
-                    self.data_array[ii]['elev'] = d_arr_copy[d_index]['elev']
-                    self.data_array[ii]['rel_east'] = d_arr_copy[d_index]['rel_east']
-                    self.data_array[ii]['rel_north'] = d_arr_copy[d_index]['rel_north']
-                    self.data_array[:]['zone'] = d_arr_copy[d_index]['zone']
-                except IndexError:
-                    print 'Could not find {0} in data_array'.format(s_key)
-            else:
-                self.data_array[ii]['station'] = mt_obj.station
-                self.data_array[ii]['lat'] = mt_obj.lat
-                self.data_array[ii]['lon'] = mt_obj.lon
-                self.data_array[ii]['east'] = mt_obj.east
-                self.data_array[ii]['north'] = mt_obj.north
-                self.data_array[ii]['elev'] = mt_obj.elev
-                try:
-                    self.data_array[ii]['rel_east'] = mt_obj.grid_east
-                    self.data_array[ii]['rel_north'] = mt_obj.grid_north
-                    rel_distance = False
-                except AttributeError:
-                    pass
-
-            # interpolate each station onto the period list
-            # check bounds of period list
-            interp_periods = self.period_list[np.where(
-                (self.period_list >= 1. / mt_obj.Z.freq.max()) &
-                (self.period_list <= 1. / mt_obj.Z.freq.min()))]
-
-            interp_z, interp_t = mt_obj.interpolate(1. / interp_periods)
-            for kk, ff in enumerate(interp_periods):
-                jj = np.where(self.period_list == ff)[0][0]
-                self.data_array[ii]['z'][jj] = interp_z.z[kk, :, :]
-                self.data_array[ii]['z_err'][jj] = interp_z.z_err[kk, :, :]
-
-                if mt_obj.Tipper.tipper is not None:
-                    self.data_array[ii]['tip'][jj] = interp_t.tipper[kk, :, :]
-                    self.data_array[ii]['tip_err'][jj] = \
-                        interp_t.tipper_err[kk, :, :]
-
-        if rel_distance is False:
-            self.get_relative_station_locations()
-
-    def _set_station_locations(self, station_obj=None):
-        """
-        take a station_locations array and populate data_array
-        """
-
-        if station_obj is not None:
-            station_locations = station_obj.station_locations
-
-        if self.data_array is None:
-            self._set_dtype((len(self.period_list), 2, 2),
-                            (len(self.period_list), 1, 2))
-            self.data_array = np.zeros(station_locations.size,
-                                       dtype=self._dtype)
-            for d_index, s_arr in enumerate(station_locations):
-                self.data_array[d_index]['lat'] = s_arr['lat']
-                self.data_array[d_index]['lon'] = s_arr['lon']
-                self.data_array[d_index]['east'] = s_arr['east']
-                self.data_array[d_index]['north'] = s_arr['north']
-                self.data_array[d_index]['elev'] = s_arr['elev']
-                self.data_array[d_index]['rel_east'] = s_arr['rel_east']
-                self.data_array[d_index]['rel_north'] = s_arr['rel_north']
-
-        else:
-            for s_arr in station_locations:
-                try:
-                    d_index = np.where(self.data_array['station'] ==
-                                       s_arr['station'])[0][0]
-                except IndexError:
-                    print 'Could not find {0} in data_array'.format(s_arr['station'])
-                    d_index = None
-
-                if d_index is not None:
-                    self.data_array[d_index]['lat'] = s_arr['lat']
-                    self.data_array[d_index]['lon'] = s_arr['lon']
-                    self.data_array[d_index]['east'] = s_arr['east']
-                    self.data_array[d_index]['north'] = s_arr['north']
-                    self.data_array[d_index]['elev'] = s_arr['elev']
-                    self.data_array[d_index]['rel_east'] = s_arr['rel_east']
-                    self.data_array[d_index]['rel_north'] = s_arr['rel_north']
-
-    def _get_station_locations(self):
-        """
-        extract station locations from data array
-        """
-        if self.data_array is None:
-            return None
-
-        station_locations = self.data_array[['station', 'lat', 'lon',
-                                             'north', 'east', 'elev',
-                                             'rel_north', 'rel_east', 'zone']]
-        station_obj = Stations(model_epsg=self.model_epsg,
-                               model_utm_zone=self.model_utm_zone)
-        station_obj.station_locations = station_locations
-
-        return station_obj
-
-    station_locations = property(_get_station_locations,
-                                 _set_station_locations,
-                                 doc="""location of stations""")
-
-    def compute_inv_error(self):
-        """
-        compute the error from the given parameters
-        """
-        # copy values over to inversion error
-        self.data_array['z_inv_err'] = self.data_array['z_err']
-        self.data_array['tip_inv_err'] = self.data_array['tip_err']
-
-        # compute relative error for tipper
-        if 'floor' in self.error_type_tipper:
-            t_index = np.where(self.data_array['tip_err'] < self.error_value_tipper)
-            self.data_array['tip_inv_err'][t_index] = self.error_value_tipper
-        elif 'abs' in self.error_type_tipper:
-            self.data_array['tip_inv_err'][:] = self.error_value_tipper
-
-        # compute error for z
-        err_value = self.error_value_z / 100.
-        for ss in range(self.data_array.shape[0]):
-            for ff in range(max([self._t_shape[0], self._z_shape[0]])):
-                d_xx = abs(self.data_array['z'][ss, ff, 0, 0])
-                d_xy = abs(self.data_array['z'][ss, ff, 0, 1])
-                d_yx = abs(self.data_array['z'][ss, ff, 1, 0])
-                d_yy = abs(self.data_array['z'][ss, ff, 1, 1])
-                d = np.array([d_xx, d_xy, d_yx, d_yy])
-                nz = np.nonzero(d)
-
-                if d.sum() == 0.0:
-                    continue
-
-                if 'egbert' in self.error_type_z:
-                    if d_xy == 0.0:
-                        d_xy = 1.0
-                    if d_yx == 0.0:
-                        d_yx = 1.0
-                    err = err_value * np.sqrt(d_xy * d_yx)
-                    if err == 1.0:
-                        err = max([d_xx, d_xy, d_yx, d_yy]) * 10
-
-                elif 'median' in self.error_type_z:
-                    err = err_value * np.median(d[nz])
-
-                elif 'mean_od' in self.error_type_z:
-                    d = np.array(d_xy, d_yx)
-                    nz = np.nonzero(d)
-                    err = err_value * np.mean(d[nz])
-
-                elif 'eigen' in self.error_type_z:
-                    d = d.reshape((2, 2))
-                    err = err_value * np.abs(np.linalg.eigvals(d)).mean()
-                    if err == 0:
-                        err = err_value * d.flatten()[nz].mean()
-
-                else:
-                    raise NameError('{0} not understood'.format(self.error_type_z))
-
-                self.data_array['z_inv_err'][ss, ff, :, :] = err
-
-        # if there is an error floor                
-        if 'floor' in self.error_type_z:
-            f_index = np.where(self.data_array['z_inv_err'] < self.data_array['z_err'])
-            self.data_array['z_inv_err'][f_index] = self.data_array['z_err'][f_index]
-
-    def write_data_file(self, save_path=None, fn_basename=None,
-                        rotation_angle=None, compute_error=True, fill=True,
-                        elevation=False):
-        """
-        write data file for ModEM
-        
-        will save file as save_path/fn_basename
-        
-        Arguments:
-        ------------
-            **save_path** : string
-                            directory path to save data file to.
-                            *default* is cwd
-                            
-            **fn_basename** : string
-                              basename to save data file as
-                              *default* is 'ModEM_Data.dat'
-                              
-            **rotation_angle** : float
-                                angle to rotate the data by assuming N = 0, 
-                                E = 90. *default* is 0.0
-                              
-        Outputs:
-        ----------
-            **data_fn** : string
-                          full path to created data file
-                          
-        :Example: :: 
-                            
-            >>> import os
-            >>> import mtpy.modeling.modem as modem
-            >>> edi_path = r"/home/mt/edi_files"
-            >>> edi_list = [os.path.join(edi_path, edi) \
-                            for edi in os.listdir(edi_path)\
-                            if edi.find('.edi') > 0]
-            >>> md = modem.Data(edi_list, period_min=.1, period_max=300,\
-                                max_num_periods=12)
-            >>> md.write_data_file(save_path=r"/home/modem/inv1")
-        """
-
-        if save_path is not None:
-            self.save_path = save_path
-        if fn_basename is not None:
-            self.data_fn = fn_basename
-
-        self.data_fn = os.path.join(self.save_path, self.data_fn)
-
-        self.get_period_list()
-
-        # rotate data if desired
-        if rotation_angle is not None:
-            self.rotation_angle = rotation_angle
-
-        # be sure to fill in data array
-        if fill:
-            self.fill_data_array()
-            # get relative station locations in grid coordinates
-            self.get_relative_station_locations()
-
-        if elevation is False:
-            self.data_array['elev'][:] = 0.0
-
-        dlines = []
-        for inv_mode in self.inv_mode_dict[self.inv_mode]:
-            if 'impedance' in inv_mode.lower():
-                dlines.append(self.get_header_string(self.error_type_z,
-                                                     self.error_value_z,
-                                                     self.rotation_angle))
-                nsta = len(np.nonzero(np.abs(self.data_array['z']).sum(axis=(1, 2, 3)))[0])
-                nper = len(np.nonzero(np.abs(self.data_array['z']).sum(axis=(0, 2, 3)))[0])
-            elif 'vertical' in inv_mode.lower():
-                dlines.append(self.get_header_string(self.error_type_tipper,
-                                                     self.error_value_tipper,
-                                                     self.rotation_angle))
-                nsta = len(np.nonzero(np.abs(self.data_array['tip']).sum(axis=(1, 2, 3)))[0])
-                nper = len(np.nonzero(np.abs(self.data_array['tip']).sum(axis=(0, 2, 3)))[0])
-            dlines.append(self.header_string)
-            dlines.append('> {0}\n'.format(inv_mode))
-
-            if inv_mode.find('Impedance') > 0:
-                dlines.append('> exp({0}i\omega t)\n'.format(self.wave_sign_impedance))
-                dlines.append('> {0}\n'.format(self.units))
-            elif inv_mode.find('Vertical') >= 0:
-                dlines.append('> exp({0}i\omega t)\n'.format(self.wave_sign_tipper))
-                dlines.append('> []\n')
-            dlines.append('> 0\n')  # oriention, need to add at some point
-            dlines.append('> {0:>10.6f} {1:>10.6f}\n'.format(
-                self.center_point.lat[0], self.center_point.lon[0]))
-            dlines.append('> {0} {1}\n'.format(nper, nsta))
-
-            if compute_error == True:
-                self.compute_inv_error()
-
-            for ss in range(self.data_array['z'].shape[0]):
-                for ff in range(self.data_array['z'].shape[1]):
-                    for comp in self.inv_comp_dict[inv_mode]:
-                        # index values for component with in the matrix
-                        z_ii, z_jj = self.comp_index_dict[comp]
-
-                        # get the correct key for data array according to comp
-                        if comp.find('z') == 0:
-                            c_key = 'z'
-                        elif comp.find('t') == 0:
-                            c_key = 'tip'
-
-                        # get the value for that compenent at that frequency
-                        zz = self.data_array[ss][c_key][ff, z_ii, z_jj]
-                        if zz.real != 0.0 and zz.imag != 0.0 and \
-                                        zz.real != 1e32 and zz.imag != 1e32:
-                            if self.formatting == '1':
-                                per = '{0:<12.5e}'.format(self.period_list[ff])
-                                sta = '{0:>7}'.format(self.data_array[ss]['station'])
-                                lat = '{0:> 9.3f}'.format(self.data_array[ss]['lat'])
-                                lon = '{0:> 9.3f}'.format(self.data_array[ss]['lon'])
-                                eas = '{0:> 12.3f}'.format(self.data_array[ss]['rel_east'])
-                                nor = '{0:> 12.3f}'.format(self.data_array[ss]['rel_north'])
-                                ele = '{0:> 12.3f}'.format(self.data_array[ss]['elev'])
-                                com = '{0:>4}'.format(comp.upper())
-                                if self.units == 'ohm':
-                                    rea = '{0:> 14.6e}'.format(zz.real / 796.)
-                                    ima = '{0:> 14.6e}'.format(zz.imag / 796.)
-                                else:
-                                    rea = '{0:> 14.6e}'.format(zz.real)
-                                    ima = '{0:> 14.6e}'.format(zz.imag)
-
-
-                            elif self.formatting == '2':
-                                per = '{0:<14.6e}'.format(self.period_list[ff])
-                                sta = '{0:<10}'.format(self.data_array[ss]['station'])
-                                lat = '{0:> 14.6f}'.format(self.data_array[ss]['lat'])
-                                lon = '{0:> 14.6f}'.format(self.data_array[ss]['lon'])
-                                eas = '{0:> 12.3f}'.format(self.data_array[ss]['rel_east'])
-                                nor = '{0:> 15.3f}'.format(self.data_array[ss]['rel_north'])
-                                ele = '{0:> 10.3f}'.format(self.data_array[ss]['elev'])
-                                com = '{0:>12}'.format(comp.upper())
-                                if self.units == 'ohm':
-                                    rea = '{0:> 17.6e}'.format(zz.real / 796.)
-                                    ima = '{0:> 17.6e}'.format(zz.imag / 796.)
-                                else:
-                                    rea = '{0:> 17.6e}'.format(zz.real)
-                                    ima = '{0:> 17.6e}'.format(zz.imag)
-
-                            # get error from inversion error
-                            abs_err = self.data_array['{0}_inv_err'.format(c_key)][ss, ff, z_ii, z_jj]
-
-                            if np.isinf(abs_err) or np.isnan(abs_err):
-                                abs_err = 10 ** (np.floor(np.log10(abs(max([float(rea),
-                                                                            float(ima)])))))
-                            abs_err = '{0:> 14.6e}'.format(abs(abs_err))
-                            # make sure that x==north, y==east, z==+down
-                            dline = ''.join([per, sta, lat, lon, nor, eas, ele,
-                                             com, rea, ima, abs_err, '\n'])
-                            dlines.append(dline)
-
-        with open(self.data_fn, 'w') as dfid:
-            dfid.writelines(dlines)
-
-        print 'Wrote ModEM data file to {0}'.format(self.data_fn)
-        return self.data_fn
-
-    def convert_ws3dinv_data_file(self, ws_data_fn, station_fn=None,
-                                  save_path=None, fn_basename=None):
-        """ 
-        convert a ws3dinv data file into ModEM format
-        
-        Arguments:
-        ------------
-            **ws_data_fn** : string
-                             full path to WS data file
-            
-            **station_fn** : string
-                             full path to station info file output by
-                             mtpy.modeling.ws3dinv. Or you can create one using
-                             mtpy.modeling.ws3dinv.WSStation
-            
-            **save_path** : string
-                            directory path to save data file to.
-                            *default* is cwd
-                            
-            **fn_basename** : string
-                              basename to save data file as
-                              *default* is 'ModEM_Data.dat'
-                              
-        Outputs:
-        -----------
-            **data_fn** : string
-                          full path to created data file
-        
-        :Example: ::
-            
-            >>> import mtpy.modeling.modem as modem
-            >>> mdr = modem.Data()
-            >>> mdr.convert_ws3dinv_data_file(r"/home/ws3dinv/inv1/WSData.dat",
-                    station_fn=r"/home/ws3dinv/inv1/WS_Station_Locations.txt")
-        """
-
-        if os.path.isfile(ws_data_fn) == False:
-            raise ws.WSInputError('Did not find {0}, check path'.format(ws_data_fn))
-
-        if save_path is not None:
-            self.save_path = save_path
-        else:
-            self.save_path = os.path.dirname(ws_data_fn)
-
-        if fn_basename is not None:
-            self.fn_basename = fn_basename
-
-        # --> get data from data file
-        wsd = ws.WSData()
-        wsd.read_data_file(ws_data_fn, station_fn=station_fn)
-
-        ns = wsd.data['station'].shape[0]
-        nf = wsd.period_list.shape[0]
-
-        self.period_list = wsd.period_list.copy()
-        self._set_dtype((nf, 2, 2), (nf, 1, 2))
-        self.data_array = np.zeros(ns, dtype=self._dtype)
-
-        # --> fill data array
-        for ii, d_arr in enumerate(wsd.data):
-            self.data_array[ii]['station'] = d_arr['station']
-            self.data_array[ii]['rel_east'] = d_arr['east']
-            self.data_array[ii]['rel_north'] = d_arr['north']
-            self.data_array[ii]['z'][:] = d_arr['z_data']
-            self.data_array[ii]['z_err'][:] = d_arr['z_data_err'].real * \
-                                              d_arr['z_err_map'].real
-            self.data_array[ii]['station'] = d_arr['station']
-            self.data_array[ii]['lat'] = 0.0
-            self.data_array[ii]['lon'] = 0.0
-            self.data_array[ii]['rel_east'] = d_arr['east']
-            self.data_array[ii]['rel_north'] = d_arr['north']
-            self.data_array[ii]['elev'] = 0.0
-
-        # need to change the inversion mode to be the same as the ws_data file
-        if self.data_array['z'].all() == 0.0:
-            if self.data_array['tip'].all() == 0.0:
-                self.inv_mode = '4'
-            else:
-                self.inv_mode = '3'
-        else:
-            if self.data_array['tip'].all() == 0.0:
-                self.inv_mode = '2'
-            else:
-                self.inv_mode = '1'
-
-        # -->write file
-        self.write_data_file()
-
-    def convert_modem_to_ws(self, data_fn=None, ws_data_fn=None,
-                            error_map=[1, 1, 1, 1]):
-        """
-        convert a ModEM data file to WS format.
-        
-        Arguments
-        -------------
-            **data_fn** : string
-                         full path to modem data file to convert
-                         
-            **ws_data_fn** : string
-                             full path to write ws format data file
-            
-            **error_map** : [zxx, zxy, zyx, zyy] floats
-                            error map that ws uses, weights for each component
-                            *default* is [1, 1, 1, 1] for equal weighting
-        Returns
-        ------------
-            **ws_data_fn** : string
-                             full path of ws data file
-                             
-            **ws_station_fn** : string 
-                                full path to ws station file
-                                
-        Example
-        -----------
-            :Convert ModEM data file to WS: ::
-                >>> import mtpy.modeling.modem as modem
-                >>> md = modem.Data()
-                >>> md.convert_modem_to_ws(data_fn=r"/home/mt/modem/data.dat")
-        """
-
-        if self.data_fn is not None:
-            self.read_data_file(data_fn)
-
-        if ws_data_fn == None:
-            save_path = os.path.dirname(self.data_fn)
-            ws_data_fn = os.path.join(save_path, 'WS_Data.dat')
-
-        else:
-            save_path = os.path.dirname(ws_data_fn)
-
-        station_info = ws.WSStation()
-        station_info.east = self.data_array['rel_east']
-        station_info.north = self.data_array['rel_north']
-        station_info.names = self.data_array['station']
-        station_info.elev = self.data_array['elev']
-        station_info.save_path = save_path
-        station_info.write_station_file()
-
-        ws_data = ws.WSData()
-        ws_data.period_list = self.period_list.copy()
-        ws_data.z_err_map = error_map
-        ws_data.z_err = 'data'
-        z_shape = (self.period_list.size, 2, 2)
-        data_dtype = [('station', '|S10'),
-                      ('east', np.float),
-                      ('north', np.float),
-                      ('z_data', (np.complex, z_shape)),
-                      ('z_data_err', (np.complex, z_shape)),
-                      ('z_err_map', (np.complex, z_shape))]
-        ws_data.data = np.zeros(self.data_array['station'].size,
-                                dtype=data_dtype)
-        ws_data.data['station'][:] = self.data_array['station']
-        ws_data.data['east'] = self.data_array['rel_east']
-        ws_data.data['north'] = self.data_array['rel_north']
-        ws_data.data['z_data'][:, :, :] = self.data_array['z']
-        ws_data.data['z_data_err'][:, :, :] = self.data_array['z_err'] * (1 + 1j)
-        ws_data.data['z_err_map'][:, :, :] = np.array([[1, 1], [1, 1]])
-
-        ws_data.write_data_file(save_path=save_path, data_fn=ws_data_fn)
-
-        return ws_data.data_fn, station_info.station_fn
-
-    def read_data_file(self, data_fn=None):
-        """
-        read ModEM data file
-        
-        Fills attributes: 
-            * data_array
-            * period_list
-            * mt_dict
-        
-        """
-
-        if data_fn is not None:
-            self.data_fn = data_fn
-            self.save_path = os.path.dirname(self.data_fn)
-            self.fn_basename = os.path.basename(self.data_fn)
-
-        if self.data_fn is None:
-            raise ModEMError('data_fn is None, enter a data file to read.')
-        elif os.path.isfile(self.data_fn) is False:
-            raise ModEMError('Could not find {0}, check path'.format(self.data_fn))
-
-        dfid = file(self.data_fn, 'r')
-        dlines = dfid.readlines()
-        dfid.close()
-
-        header_list = []
-        metadata_list = []
-        data_list = []
-        period_list = []
-        station_list = []
-        read_impedance = False
-        read_tipper = False
-        inv_list = []
-        for dline in dlines:
-            if dline.find('#') == 0:
-                header_list.append(dline.strip())
-            elif dline.find('>') == 0:
-                metadata_list.append(dline[1:].strip())
-                if dline.lower().find('ohm') > 0:
-                    self.units = 'ohm'
-                elif dline.lower().find('mv') > 0:
-                    self.units = ' [mV/km]/[nT]'
-                elif dline.lower().find('vertical') > 0:
-                    read_tipper = True
-                    read_impedance = False
-                    inv_list.append('Full_Vertical_Components')
-                elif dline.lower().find('impedance') > 0:
-                    read_impedance = True
-                    read_tipper = False
-                    inv_list.append('Full_Impedance')
-                if dline.find('exp') > 0:
-                    if read_impedance is True:
-                        self.wave_sign_impedance = dline[dline.find('(') + 1]
-                    elif read_tipper is True:
-                        self.wave_sign_tipper = dline[dline.find('(') + 1]
-                elif len(dline[1:].strip().split()) == 2:
-                    if dline.find('.') > 0:
-                        value_list = [float(value) for value in
-                                      dline[1:].strip().split()]
-
-                        self.center_point = np.recarray(1, dtype=[('station', '|S10'),
-                                                                  ('lat', np.float),
-                                                                  ('lon', np.float),
-                                                                  ('elev', np.float),
-                                                                  ('rel_east', np.float),
-                                                                  ('rel_north', np.float),
-                                                                  ('east', np.float),
-                                                                  ('north', np.float),
-                                                                  ('zone', 'S4')])
-                        self.center_point.lat = value_list[0]
-                        self.center_point.lon = value_list[1]
-
-                        ce, cn, cz = gis_tools.project_point_ll2utm(self.center_point.lat,
-                                                                    self.center_point.lon)
-
-                        self.center_point.east = ce
-                        self.center_point.north = cn
-                        self.center_point.zone = cz
-
-                    else:
-                        pass
-
-            else:
-                dline_list = dline.strip().split()
-                if len(dline_list) == 11:
-                    for ii, d_str in enumerate(dline_list):
-                        if ii != 1:
-                            try:
-                                dline_list[ii] = float(d_str.strip())
-                            except ValueError:
-                                pass
-                        # be sure the station name is a string
-                        else:
-                            dline_list[ii] = d_str.strip()
-                    period_list.append(dline_list[0])
-                    station_list.append(dline_list[1])
-
-                    data_list.append(dline_list)
-
-        # try to find rotation angle
-        h_list = header_list[0].split()
-        for hh, h_str in enumerate(h_list):
-            if h_str.find('_deg') > 0:
-                try:
-                    self._rotation_angle = float(h_str[0:h_str.find('_deg')])
-                    print ('Set rotation angle to {0:.1f} '.format(
-                        self._rotation_angle) + 'deg clockwise from N')
-                except ValueError:
-                    pass
-
-        # find inversion mode
-        for inv_key in self.inv_mode_dict.keys():
-            inv_mode_list = self.inv_mode_dict[inv_key]
-            if len(inv_mode_list) != inv_list:
-                continue
-            else:
-                tf_arr = np.zeros(len(inv_list), dtype=np.bool)
-
-                for tf, data_inv in enumerate(inv_list):
-                    if data_inv in self.inv_mode_dict[inv_key]:
-                        tf_arr[tf] = True
-
-                if np.alltrue(tf_arr) == True:
-                    self.inv_mode = inv_key
-                    break
-
-        self.period_list = np.array(sorted(set(period_list)))
-        station_list = sorted(set(station_list))
-
-        # make a period dictionary to with key as period and value as index
-        period_dict = dict([(per, ii) for ii, per in enumerate(self.period_list)])
-
-        # --> need to sort the data into a useful fashion such that each station
-        #    is an mt object
-
-        data_dict = {}
-        z_dummy = np.zeros((len(self.period_list), 2, 2), dtype='complex')
-        t_dummy = np.zeros((len(self.period_list), 1, 2), dtype='complex')
-
-        index_dict = {'zxx': (0, 0), 'zxy': (0, 1), 'zyx': (1, 0), 'zyy': (1, 1),
-                      'tx': (0, 0), 'ty': (0, 1)}
-
-        # dictionary for true false if station data (lat, lon, elev, etc)
-        # has been filled already so we don't rewrite it each time
-        tf_dict = {}
-        for station in station_list:
-            data_dict[station] = mt.MT()
-            data_dict[station].Z = mtz.Z(z_array=z_dummy.copy(),
-                                         z_err_array=z_dummy.copy().real,
-                                         freq=1. / self.period_list)
-            data_dict[station].Tipper = mtz.Tipper(tipper_array=t_dummy.copy(),
-                                                   tipper_err_array=t_dummy.copy().real,
-                                                   freq=1. / self.period_list)
-            # make sure that the station data starts out with false to fill
-            # the data later
-            tf_dict[station] = False
-
-        # fill in the data for each station
-        for dd in data_list:
-            # get the period index from the data line
-            p_index = period_dict[dd[0]]
-            # get the component index from the data line
-            ii, jj = index_dict[dd[7].lower()]
-
-            # if the station data has not been filled yet, fill it
-            if tf_dict[dd[1]] == False:
-                data_dict[dd[1]].lat = dd[2]
-                data_dict[dd[1]].lon = dd[3]
-                data_dict[dd[1]].grid_north = dd[4]
-                data_dict[dd[1]].grid_east = dd[5]
-                data_dict[dd[1]].grid_elev = dd[6]
-                data_dict[dd[1]].station = dd[1]
-                tf_dict[dd[1]] = True
-            # fill in the impedance tensor with appropriate values
-            if dd[7].find('Z') == 0:
-                z_err = dd[10]
-                if self.wave_sign_impedance == '+':
-                    z_value = dd[8] + 1j * dd[9]
-                elif self.wave_sign_impedance == '-':
-                    z_value = dd[8] - 1j * dd[9]
-
-                if self.units == 'ohm':
-                    z_value *= 796.
-                    z_err *= 796.
-
-                data_dict[dd[1]].Z.z[p_index, ii, jj] = z_value
-                data_dict[dd[1]].Z.z_err[p_index, ii, jj] = z_err
-            # fill in tipper with appropriate values
-            elif dd[7].find('T') == 0:
-                if self.wave_sign_tipper == '+':
-                    data_dict[dd[1]].Tipper.tipper[p_index, ii, jj] = dd[8] + 1j * dd[9]
-                elif self.wave_sign_tipper == '-':
-                    data_dict[dd[1]].Tipper.tipper[p_index, ii, jj] = dd[8] - 1j * dd[9]
-                data_dict[dd[1]].Tipper.tipper_err[p_index, ii, jj] = dd[10]
-
-        # make mt_dict an attribute for easier manipulation later
-        self.mt_dict = data_dict
-
-        ns = len(self.mt_dict.keys())
-        nf = len(self.period_list)
-        self._set_dtype((nf, 2, 2), (nf, 1, 2))
-        self.data_array = np.zeros(ns, dtype=self._dtype)
-
-        # Be sure to caclulate invariants and phase tensor for each station
-        for ii, s_key in enumerate(sorted(self.mt_dict.keys())):
-            mt_obj = self.mt_dict[s_key]
-
-            # self.mt_dict[s_key].zinv.compute_invariants()
-            self.mt_dict[s_key].pt.set_z_object(mt_obj.Z)
-            self.mt_dict[s_key].Tipper._compute_amp_phase()
-            self.mt_dict[s_key].Tipper._compute_mag_direction()
-
-            self.data_array[ii]['station'] = mt_obj.station
-            self.data_array[ii]['lat'] = mt_obj.lat
-            self.data_array[ii]['lon'] = mt_obj.lon
-            self.data_array[ii]['east'] = mt_obj.east
-            self.data_array[ii]['north'] = mt_obj.north
-            self.data_array[ii]['elev'] = mt_obj.grid_elev
-            self.data_array[ii]['rel_east'] = mt_obj.grid_east
-            self.data_array[ii]['rel_north'] = mt_obj.grid_north
-
-            self.data_array[ii]['z'][:] = mt_obj.Z.z
-            self.data_array[ii]['z_err'][:] = mt_obj.Z.z_err
-            self.data_array[ii]['z_inv_err'][:] = mt_obj.Z.z_err
-
-            self.data_array[ii]['tip'][:] = mt_obj.Tipper.tipper
-            self.data_array[ii]['tip_err'][:] = mt_obj.Tipper.tipper_err
-            self.data_array[ii]['tip_inv_err'][:] = mt_obj.Tipper.tipper_err
-
-    def write_vtk_station_file(self, vtk_save_path=None,
-                               vtk_fn_basename='ModEM_stations'):
-        """
-        write a vtk file for station locations.  For now this in relative 
-        coordinates.  
-        
-        Arguments:
-        -------------
-            **vtk_save_path** : string
-                                directory to save vtk file to.  
-                                *default* is Model.save_path
-            **vtk_fn_basename** : string
-                                  filename basename of vtk file
-                                  *default* is ModEM_stations, evtk will add
-                                  on the extension .vtu
-        """
-
-        if vtk_save_path is None:
-            vtk_fn = os.path.join(self.save_path, vtk_fn_basename)
-        else:
-            vtk_fn = os.path.join(vtk_save_path, vtk_fn_basename)
-
-        pointsToVTK(vtk_fn,
-                    self.station_locations.rel_north / 1000.,
-                    self.station_locations.rel_east / 1000.,
-                    self.station_locations.elev / 1000.,
-                    data={'elevation': self.station_locations.elev})
-
-        print '--> Wrote station file to {0}'.format(vtk_fn)
-        print '-' * 50
-
-    def get_parameters(self):
-        """
-        get important parameters for documentation
-        """
-
-        parameter_list = ['error_type_z',
-                          'error_value_z',
-                          'error_type_tipper',
-                          'error_value_tipper',
-                          'wave_sign_impedance',
-                          'wave_sign_tipper',
-                          'rotation_angle',
-                          'save_path']
-
-        parameter_dict = {}
-        for parameter in parameter_list:
-            key = 'data.{0}'.format(parameter)
-            parameter_dict[key] = getattr(self, parameter)
-
-        parameter_dict['data.period_min'] = self.period_list.min()
-        parameter_dict['data.period_max'] = self.period_list.max()
-        parameter_dict['data.period_num'] = self.period_list.size
-
-        parameter_dict['data.inv_mode'] = self.inv_mode_dict[self.inv_mode]
-        parameter_dict['data.num_stations'] = self.station_locations.station.size
-        parameter_dict['data.center_point_ll'] = (self.center_point.lat[0],
-                                                  self.center_point.lon[0])
-
-        parameter_dict['data.center_point_utm'] = (self.center_point.north[0],
-                                                   self.center_point.east[0],
-                                                   self.center_point.zone[0])
-        return parameter_dict
-
-    def center_stations(self, model_fn, data_fn=None):
-        """
-        Center station locations to the middle of cells, might be useful for 
-        topography.
-    
-    
-        Arguments
-        -----------
-            **data_fn** : string
-                          full path to data file
-                          
-            **model_fn** : string
-                          full path to model file
-                          
-            **new_data_fn** : string
-                             full path to new data file
-                             *default* is None, which save as 
-                             data_fn_center.dat
-                             
-        Returns
-        -----------
-            **new_data_fn** : string
-                              full path to new data file 
-        """
-
-        if data_fn is not None:
-            self.read_data_file(data_fn)
-
-        m_obj = Model()
-        m_obj.read_model_file(model_fn)
-
-        for s_arr in self.station_locations.station_locations:
-            e_index = np.where(m_obj.grid_east >= s_arr['rel_east'])[0][0] - 1
-            n_index = np.where(m_obj.grid_north >= s_arr['rel_north'])[0][0] - 1
-
-            mid_east = m_obj.grid_east[e_index:e_index + 2].mean()
-            mid_north = m_obj.grid_north[n_index:n_index + 2].mean()
-
-            s_index = np.where(self.data_array['station'] == s_arr['station'])[0][0]
-
-            self.data_array[s_index]['rel_east'] = mid_east
-            self.data_array[s_index]['rel_north'] = mid_north
-
-    def change_data_elevation(self, model_fn, data_fn=None,
-                              res_air=1e12):
-        """
-        At each station in the data file rewrite the elevation, so the station is
-        on the surface, not floating in air.
-        
-        Arguments:
-        ------------------
-            *data_fn* : string
-                        full path to a ModEM data file
-                        
-            *model_fn* : string
-                        full path to ModEM model file that has elevation 
-                        incoorporated.
-                                            
-            *new_data_fn* : string
-                            full path to new data file name.  If None, then 
-                            new file name will add _elev.dat to input filename
-                            
-            *res_air* : float
-                        resistivity of air.  Default is 1E12 Ohm-m
-        Returns:
-        -------------
-            *new_data_fn* : string
-                            full path to new data file.
-        """
-        if data_fn is not None:
-            self.read_data_file(data_fn)
-
-        m_obj = Model()
-        m_obj.read_model_file(model_fn)
-
-        s_locations = self.station_locations.station_locations.copy()
-
-        # need to subtract one because we are finding the cell next to it
-        for s_arr in s_locations:
-            e_index = np.where(m_obj.grid_east >= s_arr['rel_east'])[0][0] - 1
-            n_index = np.where(m_obj.grid_north >= s_arr['rel_north'])[0][0] - 1
-            z_index = np.where(m_obj.res_model[n_index, e_index, :] < res_air * .9)[0][0]
-            s_index = np.where(self.data_array['station'] == s_arr['station'])[0][0]
-            self.data_array[s_index]['elev'] = m_obj.grid_z[z_index]
-
-    def project_stations_on_topography(self, model_object, air_resistivity=1e12):
-        """
-        This method is used in add_topography().
-        It will Re-write the data file to change the elevation column.
-        And update covariance mask according topo elevation model.
-        :param air_resistivity:
-        :return:
-        """
-
-        sx = self.station_locations.station_locations['rel_east']
-        sy = self.station_locations.station_locations['rel_north']
-
-        # find index of each station on grid
-        station_index_x = []
-        station_index_y = []
-        for sname in self.station_locations.station_locations['station']:
-            ss = np.where(self.station_locations.station_locations['station'] == sname)[0][0]
-            # relative locations of stations
-            sx, sy = self.station_locations.station_locations['rel_east'][ss], \
-                     self.station_locations.station_locations['rel_north'][ss]
-            # indices of stations on model grid
-            sxi = np.where((sx <= model_object.grid_east[1:]) & (
-                sx > model_object.grid_east[:-1]))[0][0]
-            syi = np.where((sy <= model_object.grid_north[1:]) & (
-                sy > model_object.grid_north[:-1]))[0][0]
-
-            # first, check if there are any air cells
-            if np.any(model_object.res_model[syi, sxi] > 0.95 * air_resistivity):
-                szi = np.amin(
-                    np.where((model_object.res_model[syi, sxi] < 0.95 * air_resistivity))[0])
-            # otherwise place station at the top of the model
-            else:
-                szi = 0
-
-            # get relevant grid point elevation
-            topoval = model_object.grid_z[szi]
-
-            station_index_x.append(sxi)
-            station_index_y.append(syi)
-
-            # update elevation in station locations and data array, +1 m as 
-            # data elevation needs to be below the topography (as advised by Naser)
-            self.station_locations.station_locations['elev'][ss] = topoval + 0.1
-            self.data_array['elev'][ss] = topoval + 0.1
-            print self.data_array['elev'][ss]
-
-
-        #        logger.debug("Re-write data file after adding topo")
-        self.write_data_file(fill=False, elevation=True)  # (Xi, Yi, Zi) of each station-i may be shifted
-
-        # debug self.Data.write_data_file(save_path='/e/tmp', fill=False)
-
-        return (station_index_x, station_index_y)
 
 
 # ==============================================================================
