+++ mtpy/utils/shapefiles.py
@@ -7,6 +7,7 @@ Created on Sun Apr 13 12:32:16 2014
 
 @author: jrpeacock
 """
+import mtpy.modeling.ModEM
 from mtpy.utils.gis_tools import transform_ll_to_utm
 
 try:
@@ -377,7 +378,7 @@ class PTShapeFile(object):
 
         """
 
-        modem_obj = modem.Data()
+        modem_obj = mtpy.modeling.ModEM.Data()
         modem_obj.read_data_file(modem_data_fn)
 
         self.plot_period = modem_obj.period_list.copy()
@@ -397,7 +398,7 @@ class PTShapeFile(object):
         """
 
         # first get the data and response and place them in array for later use
-        modem_data_obj = modem.Data()
+        modem_data_obj = mtpy.modeling.ModEM.Data()
         modem_data_obj.read_data_file(modem_data_fn)
 
         self.plot_period = modem_data_obj.period_list.copy()
@@ -407,7 +408,7 @@ class PTShapeFile(object):
 
         self._set_rotation_angle(rotation_angle)
 
-        modem_resp_obj = modem.Data()
+        modem_resp_obj = mtpy.modeling.ModEM.Data()
         modem_resp_obj.read_data_file(modem_resp_fn)
 
         # rotate model response
@@ -584,7 +585,7 @@ class PTShapeFile(object):
         """
 
         # first get the data and response and place them in array for later use
-        modem_data_obj = modem.Data()
+        modem_data_obj = mtpy.modeling.ModEM.Data()
         modem_data_obj.read_data_file(modem_data_fn)
 
         self.plot_period = modem_data_obj.period_list.copy()
@@ -594,7 +595,7 @@ class PTShapeFile(object):
 
         self._set_rotation_angle(rotation_angle)
 
-        modem_resp_obj = modem.Data()
+        modem_resp_obj = mtpy.modeling.ModEM.Data()
         modem_resp_obj.read_data_file(modem_resp_fn)
 
         # rotate model response
@@ -922,7 +923,7 @@ class TipperShapeFile(object):
         for plot_per in self.plot_period:
             self.tip_dict[plot_per] = []
             for mt_obj in self.mt_obj_list:
-                mt_obj.Tipper._compute_mag_direction()
+                mt_obj.Tipper.compute_mag_direction()
                 try:
                     p_index = [ff for ff, f2 in enumerate(1. / mt_obj.Z.freq)
                                if (f2 > plot_per * (1 - self.ptol)) and
@@ -1259,7 +1260,7 @@ class TipperShapeFile(object):
 
         """
 
-        modem_obj = modem.Data()
+        modem_obj = mtpy.modeling.ModEM.Data()
         modem_obj.read_data_file(modem_data_fn)
 
         self.plot_period = modem_obj.period_list.copy()
@@ -1278,10 +1279,10 @@ class TipperShapeFile(object):
         write residual tipper files for modem
 
         """
-        modem_data_obj = modem.Data()
+        modem_data_obj = mtpy.modeling.ModEM.Data()
         modem_data_obj.read_data_file(modem_data_fn)
 
-        modem_resp_obj = modem.Data()
+        modem_resp_obj = mtpy.modeling.ModEM.Data()
         modem_resp_obj.read_data_file(modem_resp_fn)
 
         self.plot_period = modem_data_obj.period_list.copy()
