+++ tests/modeling/test_modem_inputfiles_builder.py
@@ -24,7 +24,7 @@ current error:
 import os
 
 from mtpy.modeling.modem import Model
-from mtpy.modeling.modem import Data
+from mtpy.modeling.ModEM import Data
 from mtpy.modeling.modem import Covariance
 import numpy as np
